// BOnlineStore.UI/ClientApp/src/app/base-classes/base-enums/authentication-scopes.enum.ts
export enum AuthenticationScopesEnum {
  OPENID = 'openid',
  PROFILE = 'profile',
  DEFINITIONS_FULL_PERMISSION = 'definitions_full_permission',
  PRODUCTION_FULL_PERMISSION = 'production_full_permission',
  GATEWAY_FULL_PERMISSION = 'gateway_full_permission',
  BFF_FULL_PERMISSION = 'bff_full_permission',
  IDENTITYSERVERAPI = 'IdentityServerApi',
  OFFLINE_ACCESS = 'offline_access',
}